Исправил наименование классов в соответствии с правилами нейминга
Создал для каждой ручки создай класс (СourierApiClient и OrdersApiClient) и реализовал в них шаги из тестовых методов
